feat: ViewConfiguration 프로퍼티 타입을 LayoutSize로 수정, LayoutSize 파일 삭제
